Move rest of type classes into Prelude
