+++ packages/core/package.json
@@ -36,7 +36,6 @@
       "Async",
       "Array",
       "Boolean",
-      "Bounded",
       "Branded",
       "Case",
       "Chunk",
@@ -44,7 +43,6 @@
       "Dictionary",
       "Either",
       "EitherT",
-      "Equal",
       "FreeAssociative",
       "Hash",
       "Id",
@@ -56,14 +54,12 @@
       "NonEmptyArray",
       "Option",
       "OptionT",
-      "Ord",
       "Ordering",
       "Reader",
       "ReaderT",
       "RedBlackTree",
       "RoseTree",
       "Set",
-      "Show",
       "StateT",
       "SortedSet",
       "String",
@@ -109,6 +105,7 @@
       "Prelude/AssociativeCompose",
       "Prelude/AssociativeEither",
       "Prelude/AssociativeFlatten",
+      "Prelude/Bounded",
       "Prelude/Category",
       "Prelude/ChainRec",
       "Prelude/CommutativeBoth",
@@ -120,6 +117,7 @@
       "Prelude/CovariantWithIndex",
       "Prelude/Derive",
       "Prelude/DSL",
+      "Prelude/Equal",
       "Prelude/Extend",
       "Prelude/Foldable",
       "Prelude/FoldableWithIndex",
@@ -137,12 +135,14 @@
       "Prelude/Invariant",
       "Prelude/Monad",
       "Prelude/None",
+      "Prelude/Ord",
       "Prelude/Reduce",
       "Prelude/ReduceRight",
       "Prelude/ReduceRightWithIndex",
       "Prelude/ReduceWithIndex",
       "Prelude/Selective",
       "Prelude/Separate",
+      "Prelude/Show",
       "Prelude/ForEach",
       "Prelude/ForEachWithIndex",
       "Prelude/Wiltable",
