+++ packages/core/src/Array/operations.ts
@@ -7,14 +7,14 @@ import { pipe } from "@effect-ts/system/Function"
 import type { MutableArray } from "@effect-ts/system/Mutable"
 
 import type { Either } from "../Either"
-import type { Equal } from "../Equal"
-import { makeEqual } from "../Equal"
 import type { ArrayURI } from "../Modules"
-import * as Ord from "../Ord"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import * as DSL from "../Prelude/DSL"
-import type { Show } from "../Show"
+import type { Equal } from "../Prelude/Equal"
+import { makeEqual } from "../Prelude/Equal"
+import * as Ord from "../Prelude/Ord"
+import type { Show } from "../Prelude/Show"
 import type { Identity } from "../Structure/Identity"
 import { makeIdentity } from "../Structure/Identity"
 import type { PredicateWithIndex, Separated } from "../Utils"
