+++ packages/core/src/Boolean/index.ts
@@ -7,10 +7,10 @@ import "../Operator"
  * Partially Ported from https://github.com/0x706b/principia
  */
 import * as A from "../Array"
-import * as Eq from "../Equal"
 import type { Predicate } from "../Function"
 import type { Prod, Sum } from "../Newtype"
 import { And, BooleanProd, BooleanSum, Or } from "../Newtype"
+import * as Eq from "../Prelude/Equal"
 import * as C from "../Structure/Closure"
 import * as I from "../Structure/Identity"
