+++ packages/core/src/Dictionary/operations.ts
@@ -6,15 +6,15 @@ import * as O from "@effect-ts/system/Option"
 
 import * as A from "../Array"
 import * as E from "../Either"
-import type { Equal } from "../Equal"
-import { makeEqual } from "../Equal"
 import { identity, pipe, tuple } from "../Function"
 import type { DictionaryURI } from "../Modules"
 import type { Foldable, URI } from "../Prelude"
 import * as P from "../Prelude"
 import { succeedF } from "../Prelude"
+import type { Equal } from "../Prelude/Equal"
+import { makeEqual } from "../Prelude/Equal"
 import type * as HKT from "../Prelude/HKT"
-import type { Show } from "../Show"
+import type { Show } from "../Prelude/Show"
 import type { Associative } from "../Structure/Associative"
 import type { Closure } from "../Structure/Closure"
 import type { Identity } from "../Structure/Identity"
