+++ packages/core/src/Either/operations.ts
@@ -4,13 +4,13 @@ import type { Either } from "@effect-ts/system/Either"
 import * as E from "@effect-ts/system/Either"
 import { pipe, tuple } from "@effect-ts/system/Function"
 
-import type { Equal } from "../Equal"
 import type { EitherURI } from "../Modules"
 import type { Option } from "../Option"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
 import * as DSL from "../Prelude/DSL"
-import type { Show } from "../Show"
+import type { Equal } from "../Prelude/Equal"
+import type { Show } from "../Prelude/Show"
 import type { Associative } from "../Structure/Associative"
 import { makeAssociative } from "../Structure/Associative"
 import type { Identity } from "../Structure/Identity"
