+++ packages/core/src/Modules/index.ts
@@ -6,12 +6,10 @@ import type { Effect, EffectURI } from "@effect-ts/system/Effect"
 
 import type { Array } from "../Array"
 import type { Async } from "../Async"
-import type { Bounded } from "../Bounded"
 import type { Const } from "../Const"
 import type { Dictionary } from "../Dictionary"
 import type { Layer } from "../Effect/Layer"
 import type { Either } from "../Either"
-import type { Equal } from "../Equal"
 import type { FreeAssociative } from "../FreeAssociative"
 import type { Id } from "../Id"
 import type { Ix, IxC } from "../IndexedT"
@@ -19,10 +17,12 @@ import type { IO } from "../IO"
 import type { List } from "../List"
 import type { NonEmptyArray } from "../NonEmptyArray"
 import type { Option } from "../Option"
-import type { Ord } from "../Ord"
+import type { Bounded } from "../Prelude/Bounded"
+import type { Equal } from "../Prelude/Equal"
+import type { Ord } from "../Prelude/Ord"
+import type { Show } from "../Prelude/Show"
 import type { Reader } from "../Reader"
 import type { Tree } from "../RoseTree"
-import type { Show } from "../Show"
 import type { StateIn, StateOut } from "../StateT/Classic"
 import type { Closure } from "../Structure/Closure"
 import type { Commutative } from "../Structure/Commutative"
