+++ packages/core/src/Option/operations.ts
@@ -4,15 +4,15 @@ import * as O from "@effect-ts/system/Option"
 
 import type { Either } from "../Either"
 import { left, right } from "../Either"
-import type { Equal } from "../Equal"
 import type { Predicate } from "../Function"
 import { pipe } from "../Function"
 import type { OptionURI } from "../Modules"
-import type { Ord } from "../Ord"
-import { makeOrd } from "../Ord"
 import type { URI } from "../Prelude"
 import * as P from "../Prelude"
-import type { Show } from "../Show"
+import type { Equal } from "../Prelude/Equal"
+import type { Ord } from "../Prelude/Ord"
+import { makeOrd } from "../Prelude/Ord"
+import type { Show } from "../Prelude/Show"
 import type { Associative } from "../Structure/Associative"
 import { makeAssociative } from "../Structure/Associative"
 import type { Identity } from "../Structure/Identity"
