+++ packages/core/src/Prelude/Equal/instances.ts
@@ -2,9 +2,9 @@
 
 import * as Equal from "@effect-ts/system/Equal"
 
-import type { EqualURI } from "../Modules"
-import type { URI } from "../Prelude"
-import * as P from "../Prelude"
+import type { EqualURI } from "../../Modules"
+import type { URI } from "../index"
+import * as P from "../index"
 
 /**
  * The `AssociativeBoth` instance for `Equal`.
