+++ packages/core/src/Prelude/Ord/definition.ts
@@ -1,7 +1,7 @@
 // tracing: off
 
+import type { Ordering } from "../../Ordering"
 import type { Equal } from "../Equal"
-import type { Ordering } from "../Ordering"
 
 /**
  * `Ord[A]` provides implicit evidence that values of type `A` have a total
