+++ packages/core/src/Prelude/Ord/operations.ts
@@ -3,11 +3,11 @@
 import type { Ord } from "@effect-ts/system/Ord"
 import { makeOrd } from "@effect-ts/system/Ord"
 
-import { Associative as OrderingAssociative } from "../Ordering"
-import type { Associative } from "../Structure/Associative"
-import { makeAssociative } from "../Structure/Associative/makeAssociative"
-import type { Identity } from "../Structure/Identity"
-import { makeIdentity } from "../Structure/Identity/makeIdentity"
+import { Associative as OrderingAssociative } from "../../Ordering"
+import type { Associative } from "../../Structure/Associative"
+import { makeAssociative } from "../../Structure/Associative/makeAssociative"
+import type { Identity } from "../../Structure/Identity"
+import { makeIdentity } from "../../Structure/Identity/makeIdentity"
 
 /**
  * Returns a `Associative` such that:
