+++ packages/core/src/Prelude/Show/definitions.ts
@@ -1,6 +1,6 @@
 // tracing: off
 
-export { ShowURI } from "../Modules"
+export { ShowURI } from "../../Modules"
 
 /**
  * `Show[A]` provides implicit evidence that values of type `A` have a total
