+++ packages/core/src/Prelude/Show/index.ts
@@ -1,6 +1,6 @@
 // tracing: off
 
-import "../Operator"
+import "../../Operator"
 
 export * from "./definitions"
 export * from "./operations"
