+++ packages/core/src/RoseTree/core.ts
@@ -11,8 +11,6 @@
  */
 
 import * as A from "../Array"
-import type { Equal } from "../Equal"
-import { makeEqual } from "../Equal"
 import { pipe } from "../Function"
 import * as IO from "../IO"
 import type { TreeURI } from "../Modules"
@@ -21,7 +19,9 @@ import * as P from "../Prelude"
 import { sequenceF } from "../Prelude"
 import * as DSL from "../Prelude/DSL"
 import { getApplicativeF } from "../Prelude/DSL"
-import type { Show } from "../Show"
+import type { Equal } from "../Prelude/Equal"
+import { makeEqual } from "../Prelude/Equal"
+import type { Show } from "../Prelude/Show"
 import type { Identity } from "../Structure/Identity"
 
 export declare type Forest<A> = ReadonlyArray<Tree<A>>
