+++ packages/core/src/Set/index.ts
@@ -7,12 +7,12 @@ import type { Option } from "@effect-ts/system/Option"
 import type { Separated } from "@effect-ts/system/Utils"
 
 import type { Either } from "../Either"
-import type { Equal } from "../Equal"
-import { makeEqual } from "../Equal"
 import type { Predicate, Refinement } from "../Function"
 import { identity, not } from "../Function"
-import type { Ord } from "../Ord"
-import type { Show } from "../Show"
+import type { Equal } from "../Prelude/Equal"
+import { makeEqual } from "../Prelude/Equal"
+import type { Ord } from "../Prelude/Ord"
+import type { Show } from "../Prelude/Show"
 import type { Associative } from "../Structure/Associative"
 import { makeAssociative } from "../Structure/Associative"
 import type { Identity } from "../Structure/Identity"
