+++ packages/core/src/Structure/Identity/operations.ts
@@ -1,9 +1,9 @@
 // tracing: off
 
-import type { Bounded } from "../../Bounded"
 import type { Endomorphism } from "../../Function"
 import { identity } from "../../Function"
 import type { IdentityURI } from "../../Modules"
+import type { Bounded } from "../../Prelude/Bounded"
 import type { Derive } from "../../Prelude/Derive"
 import type { URI, URIS } from "../../Prelude/HKT"
 import type { Associative } from "../Associative"
