+++ packages/core/src/index.ts
@@ -37,7 +37,7 @@ export * as Transducer from "./Effect/Stream/Transducer"
 export * as Supervisor from "./Effect/Supervisor"
 export * as Either from "./Either"
 export * as EitherT from "./EitherT"
-export * as Equal from "./Equal"
+export * as Equal from "./Prelude/Equal"
 export * as FreeAssociative from "./FreeAssociative"
 export * as Function from "./Function"
 export * as Has from "./Has"
@@ -57,7 +57,7 @@ export * as Newtype from "./Newtype"
 export * as NonEmptyArray from "./NonEmptyArray"
 export * as Option from "./Option"
 export * as OptionT from "./OptionT"
-export * as Ord from "./Ord"
+export * as Ord from "./Prelude/Ord"
 export * as Ordering from "./Ordering"
 export * as Any from "./Prelude/Any"
 export * as Applicative from "./Prelude/Applicative"
@@ -121,7 +121,7 @@ export * as Record from "./Dictionary"
 export * as RedBlackTree from "./RedBlackTree"
 export * as RoseTree from "./RoseTree"
 export * as Set from "./Set"
-export * as Show from "./Show"
+export * as Show from "./Prelude/Show"
 export * as SortedSet from "./SortedSet"
 export * as StateT from "./StateT"
 export * as ParametricStateT from "./StateT/Parametric"
