+++ packages/core/test/tree.test.ts
@@ -1,4 +1,4 @@
-import * as Eq from "../src/Equal"
+import * as Eq from "../src/Prelude/Equal"
 import * as Tree from "../src/RoseTree"
 
 const a = Tree.make("root", [Tree.make("leaf1"), Tree.make("leaf2")])
